Bump min version for grpcio-status in spark provider (#36662)
Previously we limited grpcio minimum version to stop backtracking
of `pip` from happening and we could not do it in the limits of
spark provider, becaue some google dependencies used it and
conflicted with it. This problem is now gone as we have newer
versions of google dependencies and we can not only safely move
it to spark provider but also bump it slightly higher to limit
the amount of backtracking we need to do.

Extracted from #36537

(cherry picked from commit ded01a5aba337882fb19e03c24d7736c7154fdd8)